Explicitly convert invalid plugin state to int

+++ src/fomodinstallerdialog.cpp
@@ -414,7 +414,7 @@ QString FomodInstallerDialog::toString(IPluginList::PluginStates state)
     return "Inactive";
   if (state.testFlag(IPluginList::STATE_ACTIVE))
     return "Active";
-  throw Exception(tr("invalid plugin state %1").arg(state));
+  throw Exception(tr("invalid plugin state %1").arg(state.toInt()));
 }
 
 std::pair<bool, QString>
